+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "stdlib.h": "c",
+        "main_menu.h": "c"
+    }
+}
