+++ include/library.h
@@ -1,0 +1,10 @@
+#ifndef LIBRARY_H
+#define LIBRARY_H
+
+void moveXY(int x, int y);
+void textColor(int code);
+void resetTextColor();
+void box1();
+void mainBox();
+
+#endif // LIBRARY_H
