+++ include/main_menu.h
@@ -1,0 +1,6 @@
+#ifndef MAINMENU_H
+#define MAINMENU_H
+
+void main_menu();
+
+#endif // MAINMENU_H
