--- src/filehandler.c
@@ -1,10 +1,0 @@
-#include <stdio.h>
-int main ()
-{
-  printf("\033[10;10H");
-    printf("Arafat");
-//    printf("\033[31mMasum\033[0m\n");
-
-//    printf("\033[31mMasum\033[0m\n");
-   return 0;
-}
